// src/utils/api/promotion/promotion.enum.ts
export enum PromotionVoucherSchemaType {
  MARKDOWN = "MARKDOWN",
}

export enum PromotionVoucherStatus {
  ACTIVE = "ACTIVE",
}

export enum PromotionVoucherDisabledReason {
  PROMOTION_VOUCHER_CODE_DISABLED = "PROMOTION_VOUCHER_CODE_DISABLED",
  PROMOTION_VOUCHER_CODE_REDEEMED = "PROMOTION_VOUCHER_CODE_REDEEMED",
  PROMOTION_VOUCHER_CODE_EXPIRED = "PROMOTION_VOUCHER_CODE_EXPIRED",
  PROMOTION_VOUCHER_PROMOTION_NOT_ACTIVE = "PROMOTION_VOUCHER_PROMOTION_NOT_ACTIVE",
  PROMOTION_VOUCHER_SCHEMA_DISABLED = "PROMOTION_VOUCHER_SCHEMA_DISABLED",
  PROMOTION_VOUCHER_AUDIENCE_NOT_ELIGIBLE = "PROMOTION_VOUCHER_AUDIENCE_NOT_ELIGIBLE",
  PROMOTION_VOUCHER_TARGET_NOT_ELIGIBLE = "PROMOTION_VOUCHER_TARGET_NOT_ELIGIBLE",
  PROMOTION_VOUCHER_CONDITION_NOT_MET = "PROMOTION_VOUCHER_CONDITION_NOT_MET",
  PROMOTION_VOUCHER_CONFLICT_LOST = "PROMOTION_VOUCHER_CONFLICT_LOST",
  PROMOTION_VOUCHER_NOT_APPLICABLE = "PROMOTION_VOUCHER_NOT_APPLICABLE",
}

export enum PromotionVoucherReasonCode {
  PROMOTION_VOUCHER_VALID = "PROMOTION_VOUCHER_VALID",
  PROMOTION_VOUCHER_INVALID_CODE_FORMAT = "PROMOTION_VOUCHER_INVALID_CODE_FORMAT",
  PROMOTION_VOUCHER_NOT_FOUND = "PROMOTION_VOUCHER_NOT_FOUND",
  PROMOTION_VOUCHER_CODE_DISABLED = "PROMOTION_VOUCHER_CODE_DISABLED",
  PROMOTION_VOUCHER_CODE_REDEEMED = "PROMOTION_VOUCHER_CODE_REDEEMED",
  PROMOTION_VOUCHER_CODE_EXPIRED = "PROMOTION_VOUCHER_CODE_EXPIRED",
  PROMOTION_VOUCHER_PROMOTION_NOT_ACTIVE = "PROMOTION_VOUCHER_PROMOTION_NOT_ACTIVE",
  PROMOTION_VOUCHER_SCHEMA_DISABLED = "PROMOTION_VOUCHER_SCHEMA_DISABLED",
  PROMOTION_VOUCHER_AUDIENCE_NOT_ELIGIBLE = "PROMOTION_VOUCHER_AUDIENCE_NOT_ELIGIBLE",
  PROMOTION_VOUCHER_TARGET_NOT_ELIGIBLE = "PROMOTION_VOUCHER_TARGET_NOT_ELIGIBLE",
  PROMOTION_VOUCHER_CONDITION_NOT_MET = "PROMOTION_VOUCHER_CONDITION_NOT_MET",
  PROMOTION_VOUCHER_CONFLICT_LOST = "PROMOTION_VOUCHER_CONFLICT_LOST",
  PROMOTION_VOUCHER_NOT_APPLICABLE = "PROMOTION_VOUCHER_NOT_APPLICABLE",
}

export enum PromotionVoucherGroupType {
  PROMOTION = "PROMOTION",
}